fix(new-arch): register TurboModules and align APM/Luciq specs
- Convert RNLuciqReactnativePackage to TurboReactPackage with getModule()
  + getReactModuleInfoProvider() so modules resolve under bridgeless mode;
  using extends TurboReactPackage keeps the RN CLI package-class regex
  matching (BaseReactPackage was not detected, dropping libraryName and
  omitting RNLuciqSpec_ModuleProvider from autolinking).
- Add the 9 Screen Loading methods to src/specs/NativeAPM.ts so they are
  exposed over the codegen JSI binding on new arch.
- Align NativeLuciq spec: reportScreenChange(spanId) param.
- Fix lossy double->int conversion in updateNetworkLogSnapshot.
- Minor module cleanups (imports) and README registration steps updated
  to reflect getModule()/getReactModuleInfoProvider() instead of
  createNativeModules.

+++ android/src/main/java/ai/luciq/reactlibrary/RNLuciqNetworkLoggerModule.java
@@ -214,7 +214,7 @@ public class RNLuciqNetworkLoggerModule extends NativeNetworkLoggerSpec {
 
             NetworkLogSnapshot modifiedSnapshot = null;
             if (!url.isEmpty()) {
-                modifiedSnapshot = new NetworkLogSnapshot(url, requestHeadersMap, requestBody, responseHeadersMap, responseBody, responseCode);
+                modifiedSnapshot = new NetworkLogSnapshot(url, requestHeadersMap, requestBody, responseHeadersMap, responseBody, (int) responseCode);
             } else {
                 LuciqRNLogger.d(NET_TAG, "[updateNetworkLogSnapshot] Empty URL — snapshot will be null (request filtered/removed)");
             }
